Extract collection wrappers to separate files
